feat: #39 mask 유틸함수 파일단위로 분리
