+++ src/utils/mask/index.ts
@@ -1,0 +1,2 @@
+export { maskName } from './maskName';
+export { maskPhoneNumber } from './maskPhoneNumber';
